[csc] Use grpc CatalogSources
"internal" CatalogSources uses ConfigMaps to store operator artifacts.
This has the downside of running into Etcd space limits of 1MB. To
overcome that we are moving to using "grpc" CatalogSources that are
backed by a sqlite database that is exposed using a grpc interface from
a pod. To achieve this the following changes have been made:

- Add a registry builder that creates all the necessary resources for
the grpc pod. We create a deployment with a pod spec that launches the
grpc registry container and use a service to front it. The container is
given a service account with the required roles and bindings.

- Update the operator's roles to allow it to interact with services,
deployments, roles and role bindings.

- Update the image-references to include the registry image